requirements.txt
pandas
numpy
matplotlib
seaborn
mysql-connector-python
sqlalchemy
python-dotenv
vnstock3
jupyter